% Generated by roxygen2 (4.0.2): do not edit by hand
\name{expand.grid.df}
\alias{expand.grid.df}
\title{Expand grid}
\usage{
expand.grid.df(..., unique = TRUE)
}
\arguments{
  \item{...}{list of data frames (first varies fastest)}

  \item{unique}{only use unique rows?}
}
\description{
Expand grid of data frames
}
\details{
Creates new data frame containing all combination of rows
from data.frames in ...

This is an exact copy of the \code{expand.grid.df} function
in the \code{reshape} package.  I chose to include it as a
function within this package to avoid using a dependency to
a package that is no longer maintained.
}
\examples{
expand.grid.df(data.frame(a=1,b=1:2))
expand.grid.df(data.frame(a=1,b=1:2), data.frame())
expand.grid.df(data.frame(a=1,b=1:2), data.frame(c=1:2, d=1:2))
expand.grid.df(data.frame(a=1,b=1:2), data.frame(c=1:2, d=1:2), data.frame(e=c("a","b")))
}
\author{
Hadley Wickham
}

