% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/interval_lrp.R
\name{interval_lrp}
\alias{interval_lrp}
\title{Sample Size for Likelihood Ratio Positive Confidence Intervals}
\usage{
interval_lrp(lr = NULL, n = NULL, sens, spec, weights = list(c(0.5, 0.5)),
  alpha = 0.05)
}
\arguments{
\item{lr}{The lower (or upper) limit of the likelihood ratio positive desired for the study.}

\item{n}{The total sample size for the study.  Refer to the Details for strategies
for calculating \eqn{n} from the disease or non-disease group size.}

\item{sens}{The estimated or expected sensitivity of the diagnostic test.}

\item{spec}{The estimated or expected specificity of the diagnostic test.}

\item{weights}{A list of vectors, each of length 2, providing the weights for
the diease and non-disease groups.  By default, it is
assumed that there are as many patients with the disease as there are without.
The values in each vector will be normalized (summed and divided by one),
so they do not necessarily need to sum to 1.0.  Note that while
decreasing the weight of the disease group may reduce the sample size requirement,
it will increase the width of the confidence interval around the likelihood ratio negative.}

\item{alpha}{Significance level (or 1-confidence)}
}
\value{
Returns an object of class \code{HazPwr} with subclasses \code{lrp}
 and \code{est} (actual appearance: \code{lrp_est_HazPwr}.  The object
 contains a data frame of the parameters passed to the function
 (expanded using \code{expand.grid}) and the corresponding sample size
 estimates.  Sample size estimates are provided for the diseased group, the
 non-diseased group, and the total sample size.  Specific fields are:
 \enumerate{
 \item \code{lr}{Lower limit of the confidence interval for the likelihood
   ratio positive.}
 \item \code{sens}{Sensitivity of the proposed test.}
 \item \code{spec}{Specificity of the proposed test.}
 \item \code{disease.rate}{Rate of disease in the study population.}
 \item \code{alpha}{Desired significance level.}
 \item \code{n1_est}{The estimated (decimal) sample size for the diseased group.}
 \item \code{n2_est}{The estimated (decimal) sample size for the
   non-diseased group.}
 \item \code{n_est}{The total estimated (decimal) sample size.  This is
   the sum of \code{n1_est} and \code{n2_est}.}
 \item \code{n1}{The actual (integer) sample size for the diseased group.}
 \item \code{n2}{The acutal (integer) sample size for the non-diseased
   group}
 \item \code{n}{The actual total (integer) sample size.  This is the sum
   of \code{n1} and \code{n2}.}
 }
}
\description{
The likelihood ratio positive \eqn{LR^+} is a measure comparing positive results from a
diagnostic test.  Sample size requirements may be generated given an estimate of the sensitivity,
specificity, and desired lower (or upper) limit of the confidence interval.  Unbalanced designs between
diseased and non-diseased groups may be accommodated.
}
\details{
Exactly one of the parameters \code{lr}, \code{n}, and \code{alpha} must
be passed as \code{NULL}, and that parameter will be calculated from the others.
Notice that \code{alpha} has a non-\code{NULL} default, so \code{NULL} must be
explicitly passed if you want it computed.

Let \eqn{D} indicate a patient's disease status, with \eqn{D^+} indicating the patient has the
disease and \eqn{D^-} indicating no disease.  Let \eqn{T} represent a diagnostic test where \eqn{T^+}
indicates a positive test and \eqn{T^-} indicates a negative test.  Then the sensitivity, \eqn{s_e},
is defined by \eqn{s_e = P(D^+|T^+)} and specificity, \eqn{s_p} is defined by \eqn{P(D^-|T^-)}.
The likelihood ratio positive can be defined as \eqn{s_e / (1-s_p)}.
But who are we kidding, this is nearly impossible to read in this medium.
Further details about the definitions and derivations for this function
can be reviewed by running \code{vignette('LikelihoodRatioPositive')}.

In cases where the total sample size is not known, the total sample size may
be calculated using either \eqn{n = n_h + (r/(1-r)) * n_h} or
\eqn{n_d + (1-r)/r * n_d}, where \eqn{n_h} is the size of the
non-disease group, \eqn{n_d} is the size of the disease group,
and \eqn{r} is the disease rate in the total group.  The vignette given above
justifies this approach in Section 5b.  An example is given in Section 4e.
}
\examples{
# Problem 1 From Simel Article
interval_lrp(lr=2.0, n=NULL, sens=.80, spec=.73)

# Problem 3 From Simel Article
interval_lrp(lr=2.0, n=NULL, sens=.80, spec=.73, weights=list(c(1, 5)))
}
\author{
Benjamin Nutter
}
\references{
David L. Simel, Gregory P. Samsa, and David B. Matchar, ``Likelihood Ratios with confidence: sample size estimation
for diagnostic test studies," \emph{Journal of Clinical Epidemiology}, Vol 44, No 8, pp 763-770, 1991.
}

