% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval_p1.R
\name{interval_p1}
\alias{interval_p1}
\title{Study Parameters for One Sample Z-Test of Proportions}
\usage{
interval_p1(n = NULL, E = NULL, p = 0.5, alpha = 0.05, tail = "both",
  upper = FALSE, interval_min = NULL, interval_max = NULL, ...)
}
\arguments{
\item{n}{integerish value, the sample size. Must be on the interval 
[2, Inf)}

\item{E}{numeric, the margin of error. Must be on the interval (-1, 1)}

\item{p}{numeric, the hypothesize proportion. Must be on the interval (0, 1)}

\item{alpha}{numeric, the significance level. Must be on the interval (0, 1)}

\item{tail}{character, determines the direction of the confidence interval.
Must be a subset of \code{c("both", "left", "right")}}

\item{upper}{code{logical(1)}. When \code{FALSE}, the default interval when 
solving for \code{p} is \code{c(0, 0.5)}, otherwise, it is 
\code{c(0.5, 1.0)}.}

\item{interval_min}{numeric, the minimum value of the search interval.}

\item{interval_max}{numeric, the maximum value of the search interval.}

\item{...}{additional arguments to pass to \code{uniroot}}
}
\value{
Returns a data frame with the following columns:
\itemize{
 \item{\code{n_est} } Estimated sample size.
 \item{\code{n }} Sample size rounded up to the next largest integer.
 \item{\code{p }} Hypothesized proportion
 \item{\code{E }} Margin of error
 \item{\code{alpha }} Significance level
 \item{\code{tail }} Character value indicating if the test is two-tailed,
   left-tailed, or right-tailed.
 \item{\code{np }} The product of \code{n} and \code{p}. 
 \item{\code{nq }} The product of \code{n} and \code{1 - p}.
}

\code{np} and \code{nq} are provided to assist in evaluating if the study
parameters may satisfy the assumptions of the Normal approximation to the
Binomial distribution.  Typically, these values must both be greater than
five or ten.
}
\description{
The one sample z-test of proportions may be used to design
interval estimation studies based on a normal approximation to the 
Binomial distribution.  Study parameters include the sample size, margin
of error, hypothesized proportion, and significance level.
}
\details{
Exactly one of \code{n}, \code{p}, \code{E}, and 
 \code{alpha} must be \code{NULL}.  The solution is returned for the 
 \code{NULL} value that satisfies the other arguments.
}
\section{Default Interval Limits}{

\tabular{lll}{
  Study Parameter \tab Minimum \tab Maximum \cr
  n               \tab 2       \tab 1e7     \cr
  p              \tab 0       \tab 1     \cr
  E               \tab -1      \tab 1     \cr
  alpha           \tab 0       \tab 1       
}
}

\section{Functional Requirements}{

\enumerate{
 \item When \code{n = NULL}, return the value of \code{n} that satisfies 
   the other arguments.
 \item When \code{p = NULL}, return the value of \code{p} that 
   satisfies the other arguments.
 \item When \code{E = NULL}, return the value of \code{E} that satisfies
   the other arguments.
 \item When \code{alpha = NULL}, return the value of \code{alpha} that
   satisfies the other arguments.
 \item Cast an error if the number of \code{n}, \code{p}, \code{E},
   and \code{alpha} that are \code{NULL} is not exactly one.
 \item Cast an error if \code{n} is not integerish on the interval [2, Inf).
 \item Cast an error if \code{p} is not numeric on the interval 
   (-1, 1)
 \item Cast an error if \code{E} is not numeric.
 \item Cast an error if \code{alpha} is not numeric on the interval (0, 1)
 \item Cast an error if \code{tail} is not a subset of 
   \code{c("both", "left", "right")}
 \item Cast an error if \code{interval_min} is not \code{numeric(1)}
 \item Cast an error if \code{interval_max} is not \code{numeric(1)}
 \item Cast en error if \code{upper} is not \code{logical(1)}
}
}
\examples{
# Example from https://onlinecourses.science.psu.edu/stat500/node/31
interval_p1(p = 0.72, E = 0.01, alpha = 0.05)

}
\author{
Benjamin Nutter
}

