% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval_t1.R
\name{interval_t1}
\alias{interval_t1}
\title{Estimate Study Parameters for a One Sample t-Interval}
\source{

}
\usage{
interval_t1(E = NULL, s = NULL, n = NULL, alpha = 0.05,
  two_tail = TRUE, interval_min = NULL, interval_max = NULL)
}
\arguments{
\item{E}{Margin of error (\code{E} > 0)}

\item{s}{Sample standard deviation (\code{s > 0})}

\item{n}{Sample size (\code{n > 1})}

\item{alpha}{Significance level (or 1-confidence; \code{0 < alpha < 1}). For 
convenience, it is permitted to pass values such as \code{seq(0, 1, by = 0.1)}.
Strictly speaking, \code{0} and \code{1} are not on the interval (0, 1), but 
they are removed with a message prior to the argument validation. (Note: this 
could potentially result in an error on the assertion that \code{alpha} have
at length >= 1, if, for example, \code{alpha = c(0, 1)})}

\item{two_tail}{Logical values indicating if the interval is one or two tailed.}

\item{interval_min}{The minimum value for the search interval in 
\code{uniroot}.  If \code{NULL}, a default value will be chosen based on the 
NULL study parameter.  See Default Interval Limits.}

\item{interval_max}{The maximum value for the search interval in 
\code{uniroot}.  If \code{NULL}, a default value will be chosen based on the
NULL study parameter.  See Default Interval Limits.}
}
\value{
A data frame with six columns
 \itemize{
   \item \code{n_est}: Exact sample size.  When \code{n} is not 
       \code{NULL}, this will be identical to \code{n}.
   \item \code{n}: Estimated sample size; the next largest integer 
       after \code{n_est}.
   \item \code{E}: Margin of error
   \item \code{s}: Sample Standard Deviation
   \item \code{alpha}: Desired significance level.
   \item \code{two_tail}: Logical value indicating if the test was two tailed.
 }
}
\description{
Estimate a missing study parameter for a one sample t-based 
    confidence interval for the population mean, given the remaining 
    study parameters are known. The study parameters for the one sample 
    t-based interval are \code{E} (margin of error), \code{n} (sample size),
    \code{s} (standard deviation) and \code{alpha} (significance level). 
    
    The t-based confidence interval is commonly used in estimating
    population mean using sample data.  It is based on Student's t 
    distribution, an adaptation of the standard normal distribution for
    cases when the population standard deviation is not known.
}
\details{
Exactly one of the parameters \code{n}, \code{E}, \code{s} and 
\code{alpha} must be passed as \code{NULL}, and that parameter will be 
calculated from the others.  Notice that \code{alpha} has a 
non-\code{NULL} default, so \code{NULL} must be explicitly passed 
if you want it computed.

The formula used in parameter calculation is:

\deqn{E = t(alpha, n-1) * s / \sqrt{n}}
}
\section{Default Interval Limits}{


\tabular{ccc}{
  Study Parameter \tab Lower Limit \tab Upper Limit \cr
  \code{E}        \tab \code{0}    \tab \code{1e8}  \cr
  \code{n}        \tab \code{2}    \tab \code{1e8}  \cr
  \code{s}        \tab \code{0}    \tab \code{1e8}  \cr 
  \code{alpha}    \tab \code{0}    \tab \code{1}
}

The need for using limits other than the default ought to be rare.  The only
purpose for doing so would be to reduce the computational time required
to find a solution via \code{uniroot}. But \code{uniroot} is already quite 
efficient. Be warned that no validations are performed on the user supplied 
interval limits to confirm that the limits are within the domains of each 
study parameter.
}

\section{Functional Requirements}{

\enumerate{
  \item{Accepts values of \code{E} greater than 0}
  \item{Accepts integerish values of \code{n} greater than or equal to 2}
  \item{Accepts values of \code{s} greater than 0}
  \item{Accepts values of \code{alpha} between 0 and 1}
  \item{Permit one and two tail interval calculations}
  \item{The missing study parameter is calculated correctly.}
}
}
\examples{

# Use Student's sleep data, group 1
data(sleep)

student_E <- 
  (sleep$extra[sleep$group == 1] \%>\%
  t.test \%$\%
  conf.int \%>\%
  diff) / 2
student_s <- sd(sleep$extra[sleep$group == 1])

# Calculate the margin of error 

interval_t1(E = NULL,
            s = student_s,
            n = 10,
            alpha = 0.05,
            two_tail = TRUE)
            
# Calculate sample size for two different significance levels
interval_t1(E = student_E,
            s = student_s,
            alpha = c(0.05, 0.10),
            two_tail = TRUE)
            
# Calculate sample size over a range of margins of error
# and with two significance levels
(SampleSize <- 
  interval_t1(E = seq(1.0, 2.0, by = 0.01),
              s = student_s,
              alpha = c(0.05, 0.10),
              two_tail = TRUE))
            
library(ggplot2)
ggplot(data = SampleSize,
       mapping = aes(x = E,
                     y = n_est,
                     colour = factor(alpha))) + 
  geom_line()

}
\author{
Benjamin Nutter
}

