% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval_z1.R
\name{interval_z1}
\alias{interval_z1}
\title{Study Parameters for Normal-Distribution based Confidence Intervals}
\usage{
interval_z1(n = NULL, sigma = NULL, E = NULL, alpha = 0.05,
  tail = "both", interval_min = NULL, interval_max = NULL, ...)
}
\arguments{
\item{n}{integerish value on the interval (2, Inf), the sample size.}

\item{sigma}{\code{numeric} on the interval (0, Inf), population standard 
deviation}

\item{E}{\code{numeric} on the interval (-Inf, Inf), margin of error}

\item{alpha}{\code{numeric} on the interval (0, 1), significance level.}

\item{tail}{\code{character}, a subset of \code{c("both", "left", "right")}}

\item{interval_min}{\code{numeric(1)}, the minimum value for the search
interval.}

\item{interval_max}{\code{numeric(1)}, the maximum value for the search
interval.}

\item{...}{Additional arguments to pass to \code{uniroot}.}
}
\value{
Returns a data frame with the following columns:
\itemize{
 \item{\code{n_est} } Estimated sample size.
 \item{\code{n }} Sample size rounded up to the next largest integer.
 \item{\code{sigma }} Population standard deviation
 \item{\code{E }} Margin of error
 \item{\code{alpha }} Significance level
 \item{\code{tail }} Character value indicating if the test is two-tailed,
   left-tailed, or right-tailed.
}
}
\description{
The standard normal distribution can be transformed to produce
  confidence intervals. These intervals assume the population standard 
  deviation is known.  In practice, it is considered better to use the
  t-based confidence intervals.  This function permits the user to solve
  for the sample size, margin of error, standard deviation, or 
  significance level.
}
\details{
Exactly one of \code{n}, \code{sigma}, \code{E}, and 
 \code{alpha} must be \code{NULL}.  The solution is returned for the 
 \code{NULL} value that satisfies the other arguments.
}
\section{Default Interval Limits}{

\tabular{lll}{
  Study Parameter \tab Minimum \tab Maximum \cr
  n               \tab 2       \tab 1e7     \cr
  sigma           \tab 0       \tab 1e7     \cr
  E               \tab -1e7    \tab 1e7     \cr
  alpha           \tab 0       \tab 1       
}
}

\section{Functional Requirements}{

\enumerate{
 \item When \code{n = NULL}, return the value of \code{n} that satisfies 
   the other arguments.
 \item When \code{sigma = NULL}, return the value of \code{sigma} that 
   satisfies the other arguments.
 \item When \code{E = NULL}, return the value of \code{E} that satisfies
   the other arguments.
 \item When \code{alpha = NULL}, return the value of \code{alpha} that
   satisfies the other arguments.
 \item Cast an error if the number of \code{n}, \code{sigma}, \code{E},
   and \code{alpha} that are \code{NULL} is not exactly one.
 \item Cast an error if \code{n} is not integerish on the interval [2, Inf).
 \item Cast an error if \code{sigma} is not numeric on the interval 
   (0, Inf)
 \item Cast an error if \code{E} is not numeric.
 \item Cast an error if \code{alpha} is not numeric on the interval (0, 1)
 \item Cast an error if \code{tail} is not a subset of 
   \code{c("both", "left", "right")}
 \item Cast an error if \code{interval_min} is not \code{numeric(1)}
 \item Cast an error if \code{interval_max} is not \code{numeric(1)}
}
}
\examples{
# Example from https://onlinecourses.science.psu.edu/stat200/node/256
interval_z1(sigma = 10, E = 2, alpha = 0.10)

# Example from https://www.isixsigma.com/tools-templates/sampling-data/how-determine-sample-size-determining-sample-size/
interval_z1(sigma = 6.95, E = 1, alpha = 0.05)

}
\author{
Benjamin Nutter
}

