% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sim_wilcoxon}
\alias{sim_wilcoxon}
\title{Simulate the Power of a Wilcoxon Rank Sum Test}
\usage{
sim_wilcoxon(n, ..., weights = list(c(1, 1)), alpha = 0.05, nsim = 10000,
  seed = NULL)
}
\arguments{
  \item{n}{Total sample size}

  \item{...}{two distribution functions should be entered.
  See 'Details'.}

  \item{weights}{a list of vectors giving the proposed
  weights for the two groups.  Each vector will be
  normalized.}

  \item{alpha}{Significance level for the test.}

  \item{nsim}{Number of simulations to run per set of
  conditions}

  \item{seed}{The value of the random number generator seed
  (for reproducibility)}
}
\value{
Returns a data frame with the following fields: \enumerate{
\item \code{n_total} Total sample size \item \code{n1}
Group 1 sample size \item \code{n2} Group 2 sample size
\item \code{k} The proportion of the total sample size
allotted to Group 1 \item \code{alpha} Significance level
\item \code{power} Estimate of simulated power \item
\code{nsim} Number of simulations performed \item
\code{pop1_param} Parameters for the random sampling of
group 1 \item \code{pop2_param} Parameters for the random
sampling of group 2 \item \code{pop1_dist} Random sampling
function for group 1 \item \code{pop2_dist} Random sampling
function for group 2 }
}
\description{
Generates random samples from any two specified
distributions and compares the samples by a Wilcoxon rank
sum test.  Power is calculated as the proportion of tests
that correctly reject the null hypothesis.
}
\details{
Distribution functions should be entered as a valid random
variable generating function, but excluding the first
argument.  For example, to sample from a Poisson
distribution with a mean of 3, one would enter
\code{rpois(lambda=3)}.  Multiple values may be given to an
argument to generate power under multiple conditions, such
as \code{rpois(lambda=c(3, 5, 7))}.
}
\examples{
sim_wilcoxon(n=30,
             weights=list(c(1, 1), c(1, 3)),
             rpois(lambda=c(2.1, 3.1, 4.1)),
             rpois(lambda=3.53),
             nsim=100)
}
\author{
Benjamin Nutter
}

