% Generated by roxygen2 (4.0.2): do not edit by hand
\name{test_binomial}
\alias{test_binomial}
\title{Power and Sample Size Analysis for a Binomial Test}
\usage{
test_binomial(n = NULL, p0 = NULL, p1 = NULL, power = NULL,
  alpha = 0.05, alternative = "two.tailed", conservative = FALSE,
  n_limits = c(2L, 200L))
}
\arguments{
  \item{n}{The sample size, or number of trials}

  \item{p0}{The value of the probability of a success under
  the null hypothesis}

  \item{p1}{The value of the probability of a success under
  the alternative hypothesis}

  \item{power}{The power of the test}

  \item{alpha}{Significance level for the test}

  \item{alternative}{A character vector giving the
  alternative to the test. Multiple values may be given,
  but the values must be \code{"two.tailed"},
  \code{"left.tailed"}, or \code{"right.tailed"}.}

  \item{conservative}{A logical vector.  This determines if
  the sample size selected is conservative (larger).  This
  decision is required because the power as a function of
  sample size is non-montonic in the binomial distribution.
  In practice, it is usually better to look at both and
  select the sample size where \code{alpha_actual} is
  closest to \code{alpha}.}

  \item{n_limits}{The limits of the search for when
  \code{n=NULL}.  The sample size is determined in a manner
  similar to \code{uniroot}, but \code{uniroot} doesn't
  handle discrete values.}
}
\description{
Determines the sample size, power, null proportion,
alternative proportion, or significance level for a
binomial test.  The results also return the actual power
and significance.
}
\details{
Exactly one of the parameters \code{n}, \code{p0},
\code{p1}, \code{alpha}, and \code{power} must be passed as
\code{NULL}.  The only exception is that \code{delta} may
be passed as a second \code{NULL} when \code{mu0} and
\code{mu1} are specified.

The parameters are combined via \code{expand.grid}, so all
combinations of the inputs are evaluated.
}
\examples{
#* Julia Chill's Frozen Sensations Example from O'Brien and Castelloe
test_binomial(n=c(20, 40), p0=.5, p1=.8, alpha=c(.01, .05),
              alternative='right.tailed')

#* Plot the sample size for a range of n
library(ggplot2)
Chill <- test_binomial(n=20:40, p0=.5, p1=.8, alpha=c(.01, .05),
              alternative='right.tailed')

ggplot(Chill, aes(x=n, y=power, colour=factor(alpha))) + geom_line()
}
\author{
Benjamin Nutter
}
\references{
O'Brien R, Castelloe J, "Sample-Size Analysis in Study
Planning," American Statistical Association Continuing
Education Program: Section on Teaching Statistics in the
Health Sciences, Joint Statistical Meetings, San Francisco,
CA; 5 August 2003 (Short Course Manual)

Some design choices were obtained from the r-help question
at:
\url{http://r.789695.n4.nabble.com/Sample-size-calculations-for-one-sided-binomial-exact-test-td3964313.html}
}

