% Generated by roxygen2 (4.0.2): do not edit by hand
\name{test_p2}
\alias{test_p2}
\title{Power for a Two Sample Test of Proportions Using the Normal Approximation}
\usage{
test_p2(delta = NULL, n = NULL, pbar = NULL, alpha = 0.05,
  power = NULL, delta0 = 0, weights = list(c(1, 1)), two_tail = TRUE,
  p1_null, p2_null, p1_alt, p2_alt, x1, n1, x2, n2)
}
\arguments{
  \item{delta}{The difference to be observed between the
  proportions of two groups. This may also be interpreted
  as the difference under the alternative hypothesis.}

  \item{n}{The total sample size between the two groups.
  Unequal sample sizes may be accommodated using the
  \code{weights} argument.}

  \item{alpha}{The significance level for the test.}

  \item{power}{The power of the test.}

  \item{delta0}{The difference of proportions under the
  null hypothesis.  Typically, this is set to 0 (equal
  proportions), but can be altered if desired.}

  \item{weights}{A list of weights for the two sample
  sizes.}

  \item{two_tail}{Will this be a two tailed test?  A vector
  may be given.}

  \item{p1_null,p2_null}{The proportions of p1 and p2 under
  the null hypothesis. If both of these values are given,
  their difference will replace the default value of
  \code{delta0}.}

  \item{p1_alt,p2_alt}{The proportions of p1 and p2 under
  the alternative hypothesis.  If these are both provided
  and \code{delta = NULL}, the difference of these vectors
  will be used as \code{delta}.}

  \item{x1,x2,n1,n2}{Values of the frequencies and samples
  sizes from preliminary data to be used to calculate
  \code{sd}.  If All four of these values are given and
  \code{pbar=NULL}, \code{pbar} will be calculated from the
  preliminary values.  The arguments \code{n1} and
  \code{n2} have no influence over the sample size values
  in the resulting data frame.  Note: these values are
  assumed to be integers and are quietly coerced to
  integers.}

  \item{sd}{The standard error of the difference of
  proportions.  This is the value represented by s* in the
  references.}
}
\description{
The two sample test of proportions based on the normal
approximation is a method commonly taught in introductory
statistics tests.
}

