% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/test_t1.R
\name{test_t1}
\alias{test_t1}
\title{One Sample T-Test Study Planning}
\usage{
test_t1(delta = NULL, n = NULL, sd = NULL, alpha = 0.05, power = NULL,
  two.tail = TRUE, mu0 = NA, mu1 = NA)
}
\arguments{
\item{delta}{The difference of means to be tested. This may also be
specified using \code{mu0} and \code{mu1}, if desired.  If all three
of these arguments are specified, \code{delta} overrides the other two.}

\item{n}{The sample size}

\item{sd}{The standard deviation of the difference}

\item{alpha}{The significance level}

\item{power}{The power of the test}

\item{two.tail}{Logical.  Determines if the calculations assume one or two
sided testing.  This may be a vector (ie, \code{c(TRUE, FALSE)})}

\item{mu0,mu1}{Group means for calculating \code{delta}.  The difference
  is always calculated as \code{mu1 - mu0}.  These may be specified in place
  of \code{delta} and are used to calculate \code{delta} before verifiying
  that only one of the arguments in \code{NULL}.  Thus, \code{delta} and one
  other argument may be \code{NULL} so long as these arguments are specified.
  Recycling is used if they are not the same length.}
}
\value{
Returns a data frame with the columns listed below.
\enumerate{
\item mu0 The hypothesized mean (only if \code{mu0} is not \code{NA})
\item mu1 The alternative mean (only if \code{mu1} is not \code{NA})
\item delta The difference of means
\item sd The standard deviation of the difference of means
\item alpha The significance level of the test.
\item power The power of the test
\item two.tail a logical indicating if the test assumes one or two tails
\item n_est The exact estimated sample size.  This will equal \code{n}
  whenever \code{n} is not \code{NULL}
\item n The integer sample size obtained by \code{ceiling(n_est)}.
}
}
\description{
Calculates the power, sample size, difference of means,
  standard deviation, or significance level of a one sample t-test with
  options for one and two sided testing.
}
\details{
Exactly one of the parameters \code{delta}, \code{n}, \code{sd},
  \code{alpha}, and \code{power} must be passed as \code{NULL}.  The only
  exception is that \code{delta} may be passed as a second \code{NULL} when
  \code{mu0} and \code{mu1} are specified.

  The parameters are combined via \code{expand.grid}, so all combinations
  of the inputs are evaluated.
}
\examples{
test_t1(delta=1, n=20, sd=1, two.tail=TRUE)

#* Compare to one-sample version of example from the \\code{stats} package
power.t.test(n = 20, delta = 1, type="one.sample")



#* Illustrate the multiple inputs
test_t1(delta=1, sd=1, power=0.90, two.tail=c(TRUE, FALSE))

#* Compare with the examples from stats package
power.t.test(power = .90, delta = 1, type="one.sample")
power.t.test(power = .90, delta = 1, type="one.sample", alternative = "one.sided")
}
\author{
Benjamin Nutter
}
\references{
Hogg RV, McKean JW, Craig AT, \emph{Introduction to Mathematical Statistics},
Pearson Prentice Hall 6th ed., 2005. ISBN: 0-13-008507-3
}
\seealso{
\code{power.t.test}, \code{optimize}, \code{uniroot},\cr
  \code{vignette("TTestOneSample", package="StudyPlanning")}
}

