% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_t2.R
\name{test_t2}
\alias{test_t2}
\title{Study Parameter Estimation for Two Sample T-Tests}
\usage{
test_t2(delta = NULL, n = NULL, s = NULL, alpha = 0.05, power = NULL,
  delta0 = 0, weights = list(c(1, 1)), two_tail = TRUE,
  df_type = c("default", "equal_variance", "satterthwaite"), mu1_null = NA,
  mu1_alt = NA, mu2_null = NA, mu2_alt = NA, s1 = NA, n1 = NA,
  s2 = NA, n2 = NA, var_equal = FALSE)
}
\arguments{
\item{delta}{Numeric vector of proposed observed differences between 
the two sample means.}

\item{n}{Numeric vector of the total sample size.}

\item{s}{Numeric vector of the estimated joint standard deviation between
the two samples.  Must be greater than 0.}

\item{alpha}{Numeric vector of significance levels for the test.
Must be on the interval \code{(0, 1)}.}

\item{power}{Numeric vector of statistical power for the test.  Must be 
on the interval \code{(0, 1)}.}

\item{delta0}{Numeric vector of true differences between the two population 
means.  Defaults to 0, but may be changed to propose a non-zero null
hypothesis.}

\item{weights}{A list of vectors. Each list must have length 2.  The vectors
specify the weighting ratio of sample 1 to sample 2.  \code{c(1, 1)} 
indicates equal sample sizes;  \code{c(2, 1)} indicates group 1 has 
twice as many subjects as group 2.  Proportions may also be used, but 
will be normalized to establish the weight.}

\item{two_tail}{Logical, indicating if the test should be evaluated as a
one tail or two tail test.  Only distinct values will be used, and 
permuted with each of the other study inputs.}

\item{df_type}{Character vector indicating the type of degrees 
of freedom to use.  Only one value will be used.  \code{"equal_variance"},
will result in the degrees of freedom \code{n1 + n2 - 2}. 
\code{"satterthwaite"} calculates the degrees of freedom using 
the Welch-Satterthwaite equation. The behavior of \code{"default"}
depends on the value of the \code{var_equal} argument.}

\item{mu1_null, mu2_null}{Numeric vectors giving the null hypothesis
values for the population means of groups 1 and 2, respectively.}

\item{mu1_alt, mu2_alt}{Numeric vectors giving the alternative hypothesis
values for the population means of groups 1 and 2, respectively.}

\item{s1, s2}{Numeric vectors of standard deviations for the first 
and second groups. Allows for standard deviations to be specified 
directly instead of calculating the joint standard deviation for 
\code{s}.}

\item{n1, n2}{Numeric vectors of sample sizes for the two groups. This 
allows the sample sizes to be entered directly instead of using the 
\code{weights} approach.}

\item{var_equal}{Logical vector that directs the calculation of degrees
of freedom when \code{df_type = "default"}.  When \code{var_equal = TRUE}, 
the default behavior of \code{df_type = "default"} is \code{n1 + n2 - 2}.
Otherwise, the Welch-Satterthwaite equation is used. The default value 
is \code{FALSE}}
}
\description{
The two sample t-test may be used to evaluate if a difference
  exists between two population means.  It does not assume knowledge of 
  the population standard deviations.  \code{test_t2} is similar in function
  to \code{power.t.test}, but allows greater flexibility over the 
  input values and includes support for estimating study parameters 
  with unequal sample sizes.
}
\section{Argument Groups}{

A design choice has been made to provide multiple arguments that can 
accomplish similar designs.  Thus, it is not necessary to provide a 
value for every argument to the function, but if certain arguments are 
provided values at the same time, conflicts will need to be resolved.

For example, \code{n} is used to specify the total sample size, with 
group allocation determined by \code{weights}.  Meanwhile, \code{n1} and
\code{n2} may be used to specify exact sample sizes without having to 
determine the weight allocated to each group.  If all four arguments are 
provided, a conflict exists.  The following paragraphs identify these
argument groups and describe their behavior toward each other.

\strong{Sample Size}: \code{n} and \code{weights} form a pair that describe
the size of each group.  \code{n1} and \code{n2} form another pair that 
perform the same functionality.  \code{n1} and \code{n2} are ignored so long
as both values are non-missing.  If argument is given a value, both must be
given a value, and these values will override the \code{n}/\code{weight} 
pair.

\strong{Standard Deviation}: \code{s} may be used by itself to indicate the
joint (square root of either the pooled or summed variance) standard 
deviation.  In situations where the joint value is not provided, but the
individual values are, \code{s1} and \code{s2} may be provided and the 
joint value will be calculated.  If either \code{s1} or \code{s2} has 
a value, they must both have a value. The values calculated from \code{s1}
and \code{s2} override any values passed to \code{s}.

\strong{Null Hypotheses}: \code{delta0} may be used to
indicate the value of the difference under the null hypothesis. Alternatively,
\code{mu1_null} and \code{mu2_null} may be specified.  If either one is 
specified, both must be specified, and the values calculated from those 
override any values given to \code{delta0}

\strong{Alternative Hypotheses}: \code{delta} may be used to
indicate the value of the difference under the alternative hypothesis. 
Alternatively, \code{mu1_alt} and \code{mu2_alt} may be specified.  
If either one is specified, both must be specified, and the values 
calculated from those override any values given to \code{delta}
}

\section{Building the Results Frame}{

Results are returned as a data frame.  Prior to calculating any results,
the entire data frame is constructed to hold all of the results and the
inputs.  When only the primary arguments \code{delta, n, sd, alpha, 
power, delta0} and \code{weights} are used, the results frame is a 
simple matter of \code{expand.grid}

When the alternate arguments are used, the principle is the same, but
some intermediary work is done.  There is no requirement that \code{s1} 
and \code{s2}, for example, have the same length.  In order to construct
the values that will ultimately act as \code{s}, then, \code{expand.grid} 
is used on \code{s1} and \code{s2}, returning all of the permutations 
of the two vectors, which will then permute with all of the primary 
arguments. In other words, vectors of unequal length do not use recycling, 
but use permutation.
}

\section{Functional Arguments}{

\enumerate{
  \item Exactly one of \code{delta}, \code{n}, \code{s}, \code{alpha}, or
    \code{power} must be \code{NULL}.
  \item Values for \code{delta} may be calculated from \code{mu1_alt} 
    and \code{mu2_alt}.
  \item Values for \code{n} may be calculated from \code{n1} and \code{n2}
  \item Values for \code{s} may be calculated from \code{s1} and \code{s2}
  \item Values for \code{delta0} may be calculated from \code{mu1_null}
    and \code{mu2_null}
  \item \code{alpha} strips 0 and 1 from its inputs quietly.
  \item \code{power} strips 0 and 1 from its inputs quietly.
  \item \code{two_tail} is a logical vector, but is reduced to only distinct
    values. (at most \code{TRUE} and \code{FALSE}).
  \item \code{var_equal} is a logical vector, but is reduced to only 
    distinct values (at most \code{TURE} and \code{FALSE}).
}
}

