% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_z1.R
\name{test_z1}
\alias{test_z1}
\title{One Sample Z Test of a Sample Mean}
\usage{
test_z1(delta = NULL, n = NULL, sigma = NULL, alpha = 0.05,
  power = 0.8, tail = "both", mu_null = NULL, mu_alt = NULL,
  interval_min = NULL, interval_max = NULL, ...)
}
\arguments{
\item{delta}{\code{numeric}. The magnitude of the difference between
the values of the mean under the null and alternative hypotheses.  
This may alternatively be specified by providing \code{mu_null} and 
\code{mu_alt}.}

\item{n}{\code{numeric} on the interval [2, Inf). The sample size.}

\item{sigma}{\code{numeric} on the interval (0, Inf). The population
standard deviation.}

\item{alpha}{\code{numeric} on the interval (0, 1). The significance
level of the test.}

\item{power}{\code{numeric} on the interval (0, 1). The power of the
test.}

\item{tail}{\code{character} specifying the direction of the test.
must be one of \code{c("both", "left", "right")}}

\item{mu_null}{\code{numeric} the value of the mean under the null
hypothesis. See Details.}

\item{mu_alt}{\code{numeric} the value of the mean under the alternative
hypothesis. See Details.}

\item{...}{Additional arguments to pass to \code{uniroot}}
}
\value{
Returns a data frame with the following columns:
\itemize{
 \item{\code{delta }} Estimated difference.
 \item{\code{n_est} } Estimated sample size.
 \item{\code{n }} Sample size rounded up to the next largest integer.
 \item{\code{sigma }} Population standard deviation
 \item{\code{alpha }} Significance level
 \item{\code{power }} Power of the test 
 \item{\code{tail }} Character value indicating if the test is two-tailed,
   left-tailed, or right-tailed.
}

If \code{mu_null} and \code{mu_alt} are used to calculate \code{delta}, 
the data frame will also contain columns for these values.
}
\description{
The one sample Z test of a sample mean may be used to test the
  hypothesis that an observed sample mean differs from a predetermined
  constant.  The test assumes that the data are approximately normally 
  distributed with a known population variance.
}
\details{
Exactly one of \code{delta}, \code{n}, \code{sigma}, 
  \code{alpha}, and \code{power} must be \code{NULL}.  \code{delta}
  may be a second \code{NULL} onl if \code{mu_null} and \code{mu_alt}
  are both provided. In this case, \code{delta} is calculated as 
  \code{mu_alt} - \code{mu_null}.
}
\section{Default Interval Limits}{

\tabular{lll}{
  Study Parameter \tab Minimum \tab Maximum \cr
  delta           \tab -1e7    \tab 1e7     \cr
  n               \tab 2       \tab 1e7     \cr
  sigma           \tab 0       \tab 1e7     \cr
  alpha           \tab 0       \tab 1       \cr
  power           \tab 0       \tab 1 
}
}

\section{Functional Requirements}{

\enumerate{
 \item When \code{delta = NULL}, calculate the value of \code{delta} 
   that satisfies the other requirements.
 \item When \code{n = NULL}, calculate the value of \code{n} that 
   satisfies the other requirements
 \item When \code{sigma = NULL}, calculate the value of \code{sigma}
   that satisfies the other requirements.
 \item When \code{alpha = NULL}, calculate the value of \code{alpha}
   that satisfies the other requirements.
 \item When \code{power = NULL}, calculate the value of \code{power}
   that satisfies the other requirements.
 \item Cast an error when the number of \code{delta}, \code{n},
   \code{sigma}, \code{alpha}, \code{power} is not exactly one.
 \item Calculate \code{delta} from \code{mu_null} and \code{mu_alt}
 \item Ignore \code{mu_null} and \code{mu_alt} with a warning when 
   \code{delta} is not \code{NULL}
 \item Cast an error when \code{delta = NULL} and \code{mu_null} and 
   \code{mu_alt} are not both given (if \code{delta} is the only 
   \code{NULL} study parameter).
 \item Cast an error when \code{delta} is not numeric.
 \item Cast an error when \code{n} is not numeric on the interval [2, Inf)
 \item Cast an error when \code{sigma} is not numeric on the interval 
   (0, Inf)
 \item Cast an error when \code{alpha} is not numeric on the interval 
   (0, 1)
 \item Cast an error when \code{power} is not numeric on the interval 
   (0, 1)
 \item Cast an error when \code{tail} is not a subset of 
   \code{c("both", "left", "right")}
 \item Cast an error when \code{mu_null} is not numeric.
 \item Cast an error when \code{mu_alt} is not numeric.
 \item Cast an error when \code{interval_min} is not \code{numeric(1)}.
 \item Cast an error when \code{interval_max} is not \code{numeric(1)}.
}
}
\examples{
test_z1(delta = 5, sigma = 15, n = 20, alpha = 0.05, power = NULL)

}

