% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try_uniroot.R
\name{try_uniroot}
\alias{try_uniroot}
\title{Wrapper Function for \code{uniroot} Within \code{tryCatch]}}
\usage{
try_uniroot(f, interval, ...)
}
\arguments{
\item{f}{The function for which the root is sought.}

\item{interval}{A vector containing the end-points of the interval to be 
searched for the root.}

\item{...}{Additional named or unnamed arguments to be passed to \code{f}}
}
\description{
When \code{uniroot} cannot find a solution within the given interval,
  it casts an error the causes the entire sample size estimation to fail.
  This strikes me as ungraceful, and it would be preferable to return a
  \code{NA} and allow calculations to succeed where possible.
}
\section{Functional Requirements}{

\enumerate{
  \item Return \code{NA} when \code{uniroot} casts the error 
    "f() values at end points not of opposite sign". Other errors cause 
    the function to fail. 
}
}
\seealso{
\code{\link{uniroot}}
}

