% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval_t1.R
\name{interval_t1}
\alias{interval_t1}
\title{Estimate Study Parameters for a One Sample t-Interval}
\source{

}
\usage{
interval_t1(E = NULL, n = NULL, s = NULL, alpha = 0.05,
  two_tail = TRUE, interval_min = NULL, interval_max = NULL, cl = NULL,
  ncores = NULL)
}
\arguments{
\item{E}{Margin of error (\code{E} > 0)}

\item{n}{Sample size (\code{n > 1})}

\item{s}{Sample standard deviation (\code{s > 0})}

\item{alpha}{Significance level (or 1-confidence; \code{0 < alpha < 1})}

\item{two_tail}{Logical values indicating if the interval is one or two tailed.}

\item{interval_min}{The minimum value for the search interval in 
\code{uniroot}.  If \code{NULL}, a default value will be chosen based on the 
NULL study parameter.  See Default Interval Limits.}

\item{interval_max}{The maximum value for the search interval in 
\code{uniroot}.  If \code{NULL}, a default value will be chosen based on the
NULL study parameter.  See Default Interval Limits.}

\item{cl}{A parallel computing cluster object that inherits class \code{cluster}.
When \code{NULL}, estimation is done using \code{mapply}, otherwise 
\code{clusterMap} is used.}

\item{ncores}{An integerish value designating the number of cores to use in a 
parallel cluster.  Ignored if \code{cl} is not \code{NULL}.}
}
\value{
A data frame with six columns
 \itemize{
   \item \code{n_est}: Exact sample size.  When \code{n} is not 
       \code{NULL}, this will be identical to \code{n}.
   \item \code{n}: Estimated sample size; the next largest integer 
       after \code{n_est}.
   \item \code{E}: Margin of error
   \item \code{s}: Sample Standard Deviation
   \item \code{alpha}: Desired significance level.
   \item \code{two_tail}: Logical value indicating if the test was two tailed.
 }
}
\description{
Estimate a missing study parameter for a one sample t-based 
    confidence interval for the population mean, given the remaining 
    study parameters are known. The study parameters for the one sample 
    t-based interval are \code{E} (margin of error), \code{n} (sample size),
    \code{s} (standard deviation) and \code{alpha} (significance level). 
    
    The t-based confidence interval is commonly used in estimating
    population mean using sample data.  It is based on Student's t 
    distribution, an adaptation of the standard normal distribution.
}
\details{
Exactly one of the parameters \code{n}, \code{E}, \code{s} and 
\code{alpha} must be passed as \code{NULL}, and that parameter will be 
calculated from the others.  Notice that \code{alpha} has a 
non-\code{NULL} default, so \code{NULL} must be explicitly passed 
if you want it computed.

\deqn{E = t(alpha, n-1) * s / \sqrt{n}}
}
\section{Default Interval Limits}{

\tabular{ccc}{
  Study Parameter \tab Lower Limit \tab Upper Limit \cr
  \code{E}        \tab 0           \tab 1e8         \cr
  \code{n}        \tab 2           \tab 1e8         \cr
  \code{s}        \tab 0           \tab 1e8         \cr 
  \code{alpha}    \tab 0           \tab 1
}
}

\section{Functional Requirements}{

\enumerate{
  \item{Accepts values of \code{E} greater than 0}
  \item{Remove non-positive values from \code{E} with a warning}
  \item{Cast an error if \code{E} has no positive values.}
  \item{Accepts integerish values of \code{n} greater than 1}
  \item{Remove values less than 2 from \code{n} with a warning}
  \item{Cast an error if \code{n} has no suitable values.}
  \item{Accepts values of \code{s} greater than 0}
  \item{Remove non-positive values from \code{s} with a warning}
  \item{Cast an error if \code{s} has no positive values.}
  \item{Accepts values of \code{alpha} between 0 and 1}
  \item{Remove values less than 0 or greater than 1 from 
             \code{alpha} with a warning}
  \item{Cast an error if \code{alpha} has no suitable values.}
  \item{Permit one and two tail interval calculations}
}
}
\author{
Benjamin Nutter
}

