% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_binomial.R
\name{test_binomial}
\alias{test_binomial}
\title{Power and Sample Size Analysis for a Binomial Test}
\usage{
test_binomial(n = NULL, p0 = NULL, p1 = NULL, power = NULL,
  alpha = 0.05, tail = "both", conservative = FALSE,
  interval_min = NULL, interval_max = NULL)
}
\arguments{
\item{n}{The sample size, or number of trials}

\item{p0}{The value of the probability of a success under the null hypothesis}

\item{p1}{The value of the probability of a success under the alternative hypothesis}

\item{power}{The power of the test}

\item{alpha}{Significance level for the test}

\item{conservative}{A logical vector.  This determines if the sample
size selected is conservative (larger).  This decision is required
because the power as a function of sample size is non-montonic in
the binomial distribution.  In practice, it is usually better to look
at both and select the sample size where \code{alpha_actual} is closest
to \code{alpha}.}

\item{alternative}{A character vector giving the alternative to the test.
Multiple values may be given, but the values must be \code{"two.tailed"},
\code{"left.tailed"}, or \code{"right.tailed"}.}

\item{n_limits}{The limits of the search for when \code{n=NULL}.  The
sample size is determined in a manner similar to \code{uniroot}, but
\code{uniroot} doesn't handle discrete values.}
}
\description{
Determines the sample size, power, null proportion,
  alternative proportion, or significance level for a binomial test.
  The results also return the actual power and significance.
}
\details{
Exactly one of the parameters \code{n}, \code{p0},
  \code{p1}, \code{alpha}, and \code{power} must be passed as \code{NULL}.  The only
  exception is that \code{delta} may be passed as a second \code{NULL} when
  \code{mu0} and \code{mu1} are specified.

  The parameters are combined via \code{expand.grid}, so all combinations
  of the inputs are evaluated.
}
\section{Default Interval Limits}{

\tabular{ccc}{
  Study Parameter \tab Lower Limit \tab Upper Limit \cr
  \code{n}        \tab 2           \tab 1e7         \cr
  \code{p0}       \tab 0           \tab 1           \cr
  \code{p1}       \tab 0           \tab 1           \cr
  \code{power}    \tab 0           \tab 1           \cr
  \code{alpha}    \tab 0           \tab 1
}
}

\section{Functional Requirements}{

\enumerate{
  \item When \code{n = NULL} correctly calculate the estimated sample size
    that satisfies the other arguments.
  \item When \code{p0 = NULL}, correctly calculate the value of \code{p0}
    that satisfies the other arguments.
  \item When \code{p1 = NULL}, correctly calculate the value of \code{p1}
    that satisfies the other arguments.
  \item When \code{power = NULL}, correctly calculate the value of \code{power}
    that satisfies the other arguments.
  \item When \code{alpha = NULL}, correctly calculate the value of \code{alpha}
    that satisfies the other arguments.
  \item Correctly adjust values for the chosen value of \code{alternative}
  \item Correctly adjust the values returned for the chosen value of
    \code{conservative}
  \item Cast an error if the number of \code{NULL} arguments among \code{n},
    \code{p0}, \code{p1}, \code{power}, and \code{alpha} is not 1.
  \item Cast an error if \code{n} is not integerish.
  \item Cast an error if \code{p0} is not numeric on the interval (0, 1)
  \item Cast an error if \code{p1} is not numeric on the interval (0, 1)
  \item Cast an error if \code{power} is not numeric on the interval (0, 1)
  \item Cast an error if \code{alpha} is not numeric on the interval (0, 1)
  \item Cast an error if \code{tail} is not one of \code{c("both", "left", "right")}
  \item Cast an error if \code{conservative} is not logical.
  \item Retain only unique values of \code{conservative}.
  \item Cast an error if \code{interval_min} is not \code{numeric(1)}.
  \item Cast an error if \code{interval_max} is not \code{numeric(1)}.
}
}
\examples{
#* Julia Chill's Frozen Sensations Example from O'Brien and Castelloe
test_binomial(n=c(20, 40), p0=.5, p1=.8, alpha=c(.01, .05),
              tail='right')

#* Plot the sample size for a range of n
library(ggplot2)
Chill <- test_binomial(n=20:40, p0=.5, p1=.8, alpha=c(.01, .05),
              tail='right.tailed')

ggplot(Chill, aes(x=n, y=power, colour=factor(alpha))) + geom_line()

}
\author{
Benjamin Nutter
}
\references{
O'Brien R, Castelloe J, "Sample-Size Analysis in Study Planning,"
  American Statistical Association Continuing Education Program: Section on
  Teaching Statistics in the Health Sciences, Joint Statistical Meetings,
  San Francisco, CA; 5 August 2003 (Short Course Manual)

Some design choices were obtained from the r-help question at:
\url{http://r.789695.n4.nabble.com/Sample-size-calculations-for-one-sided-binomial-exact-test-td3964313.html}
}

